+++ src/modules/employee/repositories/common.ts
@@ -1,0 +1,8 @@
+import { prisma } from "@/lib/prisma";
+import { Prisma } from "@prisma/client";
+
+export type PrismaClientOrTx = Prisma.TransactionClient | typeof prisma;
+
+export function getClient(tx?: PrismaClientOrTx): PrismaClientOrTx {
+  return tx || prisma;
+}